Refactor get_summary
