+++ scripts/__init__.py
@@ -1,0 +1,3 @@
+from .benchmark_stats import BenchmarkStats
+
+__all__ = ["BenchmarkStats"]
