+++ scripts/get_summary.py
@@ -1,57 +1,32 @@
 from csv import DictReader
 from datetime import datetime, timedelta
 from glob import glob
+from typing import List, NamedTuple
 import os
-import statistics
 import sys
 
-from scripts.get_stats import get_stats, BenchmarkStats
+from scripts import BenchmarkStats
+
+
+class Result(NamedTuple):
+    date: str
+    statistics: List[str]
+
+    def __str__(self):
+        return f"{self.date}\n{self.statistics}\n"
 
 
 def get_results(folders, number_of_days=None):
-    results_list = []
-
     from_date = (
         (datetime.utcnow() - timedelta(days=number_of_days)) if number_of_days else None
     )
+
     for folder in folders:
         date = folder.split("/")[-1].split("T")[0]
         if from_date and datetime.strptime(date, "%Y-%m-%d") < from_date:
             continue
-        result = BenchmarkStats()
-        stats = get_stats(folder)
-        result.total_requests += stats.total_requests
-        result.total_failures += stats.total_failures
-        result.average_get = statistics.mean(stats.get)
-        result.average_post = statistics.mean(stats.post)
-        result.average_total = statistics.mean(stats.get + stats.post)
-        result.error_percentage = (result.total_failures * 100) / result.total_requests
-
-        get_request_response_times = []
-        post_request_response_times = []
-
-        for file in glob(folder + '/*stats.csv'):
-
-            with open(file) as f:
-                reader = DictReader(f, delimiter=",")
-
-                get_values = []
-                post_values = []
-
-                for row in reader:
-                    percentile_99th = int(row["99%"])
-                    if '/questionnaire' in row['Name']:
-                        if row['Type'] == 'GET':
-                            get_values.append(percentile_99th)
-                        elif row['Type'] == 'POST':
-                            post_values.append(percentile_99th)
-
-            get_request_response_times.extend(get_values)
-            post_request_response_times.extend(post_values)
-
-        results_list.append([date, result])
-
-    return results_list
+        results = BenchmarkStats(folder)
+        yield Result(date, results)
 
 
 def parse_environment_variables():
@@ -81,15 +56,12 @@ def parse_environment_variables():
 if __name__ == "__main__":
     parsed_variables = parse_environment_variables()
     date_to_output = parsed_variables["output_date"]
-
     sorted_folders = sorted(glob(f"{parsed_variables['output_dir']}/*"))
-
     results = get_results(sorted_folders)
 
     for result in results:
-        if date_to_output:
-            if result[0] == date_to_output:
-                print(result[1])
-                break
+        if date_to_output and result.date == date_to_output:
+            print(result.statistics)
+            break
         else:
-            print(f"{result[0]}\n" f"{result[1]}\n")
+            print(result)
